+++ packages/x-tree-view/src/internals/utils/tree.ts
